requirements.txt
africastalking==1.2.9
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
coverage==7.6.4
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==5.1.3
django-cors-headers==4.6.0
django-oauth-toolkit==3.0.1
django-oidc-provider==0.8.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
future==1.0.0
idna==3.10
jwcrypto==1.5.6
oauthlib==3.2.2
pycparser==2.22
pycryptodomex==3.21.0
pyjwkest==1.4.2
PyJWT==2.9.0
python-dotenv==1.0.1
python-social-auth==0.3.6
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
schema==0.7.7
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
